chore: rename folder, error message
